backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.4,<4.5
psycopg[binary]>=3.2,<3.3
Pillow>=10.4,<10.5
gunicorn>=23.0,<24.0
python-decouple>=3.8,<3.9
whitenoise>=6.6,<6.7
dj-database-url>=2.1,<2.2
